refactor: remove unused store files and update import paths across the core module
