--- apps/web/ce/store/issue/helpers/filter-utils.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) 2023-present Plane Software, Inc. and contributors
- * SPDX-License-Identifier: AGPL-3.0-only
- * See the LICENSE file for details.
- */
-
-import type { IIssueDisplayFilterOptions } from "@plane/types";
-
-export const getEnabledDisplayFilters = (displayFilters: IIssueDisplayFilterOptions) => displayFilters;
