+++ apps/web/core/hooks/store/estimates/use-estimate.ts
@@ -8,7 +8,7 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // mobx store
-import type { IEstimate } from "@/plane-web/store/estimates/estimate";
+import type { IEstimate } from "@/store/estimates/estimate";
 
 export const useEstimate = (estimateId: string | undefined): IEstimate => {
   const context = useContext(StoreContext);
