+++ apps/web/core/hooks/store/user/user-permissions.ts
@@ -8,7 +8,7 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // plane web imports
-import type { IUserPermissionStore } from "@/plane-web/store/user/permission.store";
+import type { IUserPermissionStore } from "@/store/user/base-permissions.store";
 
 export const useUserPermissions = (): IUserPermissionStore => {
   const context = useContext(StoreContext);
