+++ apps/web/core/store/base-command-palette.store.ts
@@ -4,7 +4,7 @@
  * See the LICENSE file for details.
  */
 
-import { observable, action, makeObservable, runInAction } from "mobx";
+import { observable, action, computed, makeObservable, runInAction } from "mobx";
 import { computedFn } from "mobx-utils";
 // plane imports
 import type { TCreateModalStoreTypes, TCreatePageModal } from "@plane/constants";
@@ -20,6 +20,8 @@ export interface ModalData {
 }
 
 export interface IBaseCommandPaletteStore {
+  // computed
+  isAnyModalOpen: boolean;
   // observables
   isCreateProjectModalOpen: boolean;
   isCreateCycleModalOpen: boolean;
@@ -52,7 +54,7 @@ export interface IBaseCommandPaletteStore {
   toggleProfileSettingsModal: (value: { activeTab?: TProfileSettingsTabs | null; isOpen?: boolean }) => void;
 }
 
-export abstract class BaseCommandPaletteStore implements IBaseCommandPaletteStore {
+export class BaseCommandPaletteStore implements IBaseCommandPaletteStore {
   // observables
   isCreateProjectModalOpen: boolean = false;
   isCreateCycleModalOpen: boolean = false;
@@ -99,7 +101,12 @@ export abstract class BaseCommandPaletteStore implements IBaseCommandPaletteStor
       toggleAllStickiesModal: action,
       toggleProjectListOpen: action,
       toggleProfileSettingsModal: action,
+      isAnyModalOpen: computed,
     });
+  }
+
+  get isAnyModalOpen(): boolean {
+    return Boolean(this.getCoreModalsState());
   }
 
   /**
@@ -277,3 +284,7 @@ export abstract class BaseCommandPaletteStore implements IBaseCommandPaletteStor
     });
   };
 }
+
+// Aliases so consumers can keep using CommandPaletteStore / ICommandPaletteStore
+export type ICommandPaletteStore = IBaseCommandPaletteStore;
+export { BaseCommandPaletteStore as CommandPaletteStore };
