+++ apps/web/core/store/issue/helpers/issue-filter-helper.store.ts
@@ -30,7 +30,6 @@ import { EIssueLayoutTypes } from "@plane/types";
 import { getComputedDisplayFilters, getComputedDisplayProperties } from "@plane/utils";
 // lib
 import { storage } from "@/lib/local-storage";
-import { getEnabledDisplayFilters } from "@/plane-web/store/issue/helpers/filter-utils";
 
 interface ILocalStoreIssueFilters {
   key: EIssuesStoreType;
@@ -68,6 +67,7 @@ export interface IIssueFilterHelperStore {
 }
 
 export class IssueFilterHelperStore implements IIssueFilterHelperStore {
+  // oxlint-disable-next-line no-useless-constructor
   constructor() {}
 
   /**
@@ -185,7 +185,7 @@ export class IssueFilterHelperStore implements IIssueFilterHelperStore {
     defaultValues?: IIssueDisplayFilterOptions
   ): IIssueDisplayFilterOptions => {
     const computedFilters = getComputedDisplayFilters(displayFilters, defaultValues);
-    return getEnabledDisplayFilters(computedFilters);
+    return computedFilters;
   };
 
   /**
