+++ apps/web/core/store/issue/issue-details/root.store.ts
@@ -18,12 +18,12 @@ import type {
   TWorkItemWidgets,
 } from "@plane/types";
 // plane web store
-import { IssueActivityStore } from "@/plane-web/store/issue/issue-details/activity.store";
+import { IssueActivityStore } from "./activity.store";
 import type {
   IIssueActivityStore,
   IIssueActivityStoreActions,
   TActivityLoader,
-} from "@/plane-web/store/issue/issue-details/activity.store";
+} from "./activity.store";
 import type { IIssueRootStore } from "../root.store";
 import { IssueAttachmentStore } from "./attachment.store";
 import type { IIssueAttachmentStore, IIssueAttachmentStoreActions } from "./attachment.store";
