+++ apps/web/core/store/user/base-permissions.store.ts
@@ -64,7 +64,7 @@ export interface IBaseUserPermissionStore {
  * @description This store is used to handle permission layer for the currently logged user.
  * It manages workspace and project level permissions, roles and access control.
  */
-export abstract class BaseUserPermissionStore implements IBaseUserPermissionStore {
+export class BaseUserPermissionStore implements IBaseUserPermissionStore {
   loader: boolean = false;
   // constants
   workspaceUserInfo: Record<string, IWorkspaceMemberMe> = {};
@@ -150,10 +150,10 @@ export abstract class BaseUserPermissionStore implements IBaseUserPermissionStor
    * @param { string } projectId
    * @returns { EUserPermissions | undefined }
    */
-  abstract getProjectRoleByWorkspaceSlugAndProjectId: (
-    workspaceSlug: string,
-    projectId?: string
-  ) => EUserPermissions | undefined;
+  getProjectRoleByWorkspaceSlugAndProjectId = computedFn(
+    (workspaceSlug: string, projectId?: string): EUserPermissions | undefined =>
+      this.getProjectRole(workspaceSlug, projectId)
+  );
 
   /**
    * @description Fetches project-level entities that are not automatically loaded by the project wrapper.
@@ -162,7 +162,9 @@ export abstract class BaseUserPermissionStore implements IBaseUserPermissionStor
    * @param { string } projectId
    * @returns { Promise<void> }
    */
-  abstract fetchWorkspaceLevelProjectEntities: (workspaceSlug: string, projectId: string) => void;
+  fetchWorkspaceLevelProjectEntities = (workspaceSlug: string, projectId: string): void => {
+    void this.store.projectRoot.project.fetchProjectDetails(workspaceSlug, projectId);
+  };
 
   /**
    * @description Returns whether the user has the permission to access a page
@@ -353,3 +355,7 @@ export abstract class BaseUserPermissionStore implements IBaseUserPermissionStor
     }
   };
 }
+
+// Aliases so consumers can keep using UserPermissionStore / IUserPermissionStore
+export type IUserPermissionStore = IBaseUserPermissionStore;
+export { BaseUserPermissionStore as UserPermissionStore };
